Make packages/chain/consensus independent of test packages.
