+++ packages/chain/consensus/consensus.go
@@ -6,7 +6,6 @@ import (
 	"github.com/iotaledger/wasp/packages/hashing"
 
 	"github.com/iotaledger/wasp/packages/vm"
-	"github.com/iotaledger/wasp/packages/vm/runvm"
 
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/wasp/packages/state"
@@ -69,13 +68,20 @@ type workflowFlags struct {
 
 var _ chain.Consensus = &consensus{}
 
-func New(chainCore chain.ChainCore, mempool chain.Mempool, committee chain.Committee, nodeConn chain.NodeConnection, log *logger.Logger) *consensus {
+func New(
+	chainCore chain.ChainCore,
+	mempool chain.Mempool,
+	committee chain.Committee,
+	nodeConn chain.NodeConnection,
+	vmRunner vm.VMRunner,
+	log *logger.Logger,
+) *consensus {
 	ret := &consensus{
 		chain:                      chainCore,
 		committee:                  committee,
 		mempool:                    mempool,
 		nodeConn:                   nodeConn,
-		vmRunner:                   runvm.NewVMRunner(),
+		vmRunner:                   vmRunner,
 		resultSignatures:           make([]*chain.SignedResultMsg, committee.Size()),
 		log:                        log.Named("c"),
 		eventStateTransitionMsgCh:  make(chan *chain.StateTransitionMsg),
